Split off the enums from engine.rs
